+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 - Faucet now supports the usage of a remote transaction prover (#830).
 - [BUGFIX] Prevents duplicated note IDs (#842).
 - Added a required Proof-of-Work in the faucet to request tokens (#831).
-- Block producer now emits network note information (#805).
+- Added an optional API key request parameter to skip PoW in faucet (#839).
+- Block producer now emits network note information (#833).
 - Introduced Network Transaction Builder (#840).
 - Added way of executing and proving network transactions (#841).
 
@@ -33,7 +34,7 @@
 ### Fixes
 
 - Faucet times out during high load (#819).
- 
+
 ## v0.8.0 (2025-03-26)
 
 ### Enhancements
@@ -55,7 +56,6 @@
 - Optimized database by adding missing indexes (#728).
 - Added support for `Content-type` header in `get_tokens` endpoint of the faucet (#754).
 - Block frequency is now configurable (#750).
-- Batch frequency is now configurable (#750).
 
 ### Changes